# Backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisterResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/verify-phone": {
      "post": {
        "operationId": "AuthController_verifyPhone",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyPhoneDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerifyPhoneResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/otp/resend": {
      "post": {
        "operationId": "AuthController_resend",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResendOtpDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OtpSentResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenPairDto"
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/refresh": {
      "post": {
        "operationId": "AuthController_refresh",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenPairDto"
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/password/forgot": {
      "post": {
        "operationId": "AuthController_forgot",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForgotPasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OtpSentResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/password/reset": {
      "post": {
        "operationId": "AuthController_reset",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResetPasswordResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/logout": {
      "post": {
        "operationId": "AuthController_logout",
        "parameters": [],
        "responses": {
          "204": {
            "description": "Session revoked"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/sessions": {
      "get": {
        "operationId": "AuthController_sessions",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SessionDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/sessions/{id}": {
      "delete": {
        "operationId": "AuthController_revokeSession",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Session revoked"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/pin": {
      "post": {
        "operationId": "AuthController_setPin",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetPinDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PinSetResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "auth"
        ]
      },
      "put": {
        "operationId": "AuthController_changePin",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePinDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PinSetResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "auth"
        ]
      }
    },
...
    "/v1/kyc/me": {
      "get": {
        "operationId": "KycController_me",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KycMeDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Current tier, its limits, and recent submissions.",
        "tags": [
          "kyc"
        ]
      }
    },
    "/v1/kyc/tiers": {
      "get": {
        "operationId": "KycController_tiers",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TierLimitDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Limits for every tier (for \"upgrade to unlock\" UX).",
        "tags": [
          "kyc"
        ]
      }
    },
    "/v1/kyc/submissions": {
      "post": {
        "operationId": "KycController_submit",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitKycDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KycSubmissionDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "kyc"
        ]
      }
    },
    "/v1/admin/kyc/submissions": {
      "get": {
        "operationId": "AdminKycController_list",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "PENDING",
                "APPROVED",
                "REJECTED"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/KycSubmissionDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin / kyc"
        ]
      }
    },
    "/v1/admin/kyc/submissions/{id}/approve": {
      "post": {
        "operationId": "AdminKycController_approve",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApproveKycDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KycSubmissionDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin / kyc"
        ]
      }
    },
    "/v1/admin/kyc/submissions/{id}/reject": {
      "post": {
        "operationId": "AdminKycController_reject",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RejectKycDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KycSubmissionDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin / kyc"
        ]
      }
    },
    "/v1/admin/kyc/users/{userId}/audit": {
      "get": {
        "operationId": "AdminKycController_audit",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/KycAuditLogDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin / kyc"
        ]
      }
    },
...
  },
  "info": {
    "title": "PayCore API",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "RegisterDto": {
        "type": "object",
        "properties": {
          "phone": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "fullName": {
            "type": "string",
            "minLength": 2,
            "maxLength": 120
          },
          "role": {
            "type": "string",
            "description": "ADMIN accounts cannot self-register.",
            "enum": [
              "CONSUMER",
              "MERCHANT"
            ]
          }
        },
        "required": [
          "phone",
          "password",
          "fullName"
        ]
      },
      "RegisterResponseDto": {
        "type": "object",
        "properties": {
          "devOtp": {
            "type": "string",
            "description": "Only when OTP_DEV_ECHO=true (never in production)"
          },
          "userId": {
            "type": "string"
          },
          "otpExpiresAt": {
            "type": "string"
          }
        },
        "required": [
          "userId",
          "otpExpiresAt"
        ]
      },
      "VerifyPhoneDto": {
        "type": "object",
        "properties": {
          "phone": {
            "type": "string"
          },
          "code": {
            "type": "string",
            "pattern": "^\\d{6}$"
          }
        },
        "required": [
          "phone",
          "code"
        ]
      },
      "VerifyPhoneResponseDto": {
        "type": "object",
        "properties": {
          "verified": {
            "type": "boolean"
          }
        },
        "required": [
          "verified"
        ]
      },
      "ResendOtpDto": {
        "type": "object",
        "properties": {
          "phone": {
            "type": "string"
          }
        },
        "required": [
          "phone"
        ]
      },
      "OtpSentResponseDto": {
        "type": "object",
        "properties": {
          "otpExpiresAt": {
            "type": "string"
          },
          "devOtp": {
            "type": "string",
            "description": "Only when OTP_DEV_ECHO=true (never in production)"
          },
          "sent": {
            "type": "boolean"
          }
        },
        "required": [
          "sent"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "phone": {
            "type": "string"
          },
          "password": {
            "type": "string",
            "minLength": 1,
            "maxLength": 128
          },
          "deviceId": {
            "type": "string",
            "description": "Stable client-generated device identifier.",
            "minLength": 1,
            "maxLength": 128
          },
          "deviceName": {
            "type": "string",
            "minLength": 1,
            "maxLength": 128
          }
        },
        "required": [
          "phone",
          "password",
          "deviceId"
        ]
      },
      "TokenPairDto": {
        "type": "object",
        "properties": {
          "tokenType": {
            "type": "string",
            "enum": [
              "Bearer"
            ]
          },
          "accessToken": {
            "type": "string"
          },
          "accessTokenExpiresIn": {
            "type": "number",
            "description": "seconds"
          },
          "refreshToken": {
            "type": "string"
          },
          "refreshTokenExpiresAt": {
            "type": "string"
          },
          "sessionId": {
            "type": "string"
          }
        },
        "required": [
          "tokenType",
          "accessToken",
          "accessTokenExpiresIn",
          "refreshToken",
          "refreshTokenExpiresAt",
          "sessionId"
        ]
      },
      "RefreshDto": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string",
            "minLength": 20,
            "maxLength": 512
          }
        },
        "required": [
          "refreshToken"
        ]
      },
      "ForgotPasswordDto": {
        "type": "object",
        "properties": {
          "phone": {
            "type": "string"
          }
        },
        "required": [
          "phone"
        ]
      },
      "ResetPasswordDto": {
        "type": "object",
        "properties": {
          "phone": {
            "type": "string"
          },
          "code": {
            "type": "string",
            "pattern": "^\\d{6}$"
          },
          "newPassword": {
            "type": "string"
          }
        },
        "required": [
          "phone",
          "code",
          "newPassword"
        ]
      },
      "ResetPasswordResponseDto": {
        "type": "object",
        "properties": {
          "reset": {
            "type": "boolean"
          }
        },
        "required": [
          "reset"
        ]
      },
      "SessionDto": {
        "type": "object",
        "properties": {
          "deviceName": {
            "type": "string",
            "nullable": true
          },
          "userAgent": {
            "type": "string",
            "nullable": true
          },
          "ipAddress": {
            "type": "string",
            "nullable": true
          },
          "id": {
            "type": "string"
          },
          "deviceId": {
            "type": "string"
          },
          "createdAt": {
            "type": "string"
          },
          "lastUsedAt": {
            "type": "string"
          },
          "current": {
            "type": "boolean"
          }
        },
        "required": [
          "deviceName",
          "userAgent",
          "ipAddress",
          "id",
          "deviceId",
          "createdAt",
          "lastUsedAt",
          "current"
        ]
      },
      "SetPinDto": {
        "type": "object",
        "properties": {
          "password": {
            "type": "string",
            "description": "Current account password, to prove possession before setting a PIN.",
            "minLength": 1,
            "maxLength": 128
          },
          "pin": {
            "type": "string"
          }
        },
        "required": [
          "password",
          "pin"
        ]
      },
      "PinSetResponseDto": {
        "type": "object",
        "properties": {
          "pinSet": {
            "type": "boolean"
          }
        },
        "required": [
          "pinSet"
        ]
      },
      "ChangePinDto": {
        "type": "object",
        "properties": {
          "currentPin": {
            "type": "string"
          },
          "newPin": {
            "type": "string"
          }
        },
        "required": [
          "currentPin",
          "newPin"
        ]
      },
...
      "KycTier": {
        "type": "string",
        "enum": [
          "TIER_0",
          "TIER_1",
          "TIER_2",
          "TIER_3"
        ]
      },
...
      "Currency": {
        "type": "string",
        "enum": [
          "PKR",
          "AED",
          "USD"
        ],
        "description": "Currencies the user can receive in (active wallets)"
      },
...
      "TierLimitDto": {
        "type": "object",
        "properties": {
          "tier": {
            "allOf": [
              {
                "$ref": "#/components/schemas/KycTier"
              }
            ]
          },
          "currency": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Currency"
              }
            ]
          },
          "permitted": {
            "type": "boolean"
          },
          "perTransaction": {
            "$ref": "#/components/schemas/MoneyDto"
          },
          "daily": {
            "$ref": "#/components/schemas/MoneyDto"
          },
          "monthly": {
            "$ref": "#/components/schemas/MoneyDto"
          },
          "maxBalance": {
            "$ref": "#/components/schemas/MoneyDto"
          }
        },
        "required": [
          "tier",
          "currency",
          "permitted",
          "perTransaction",
          "daily",
          "monthly",
          "maxBalance"
        ]
      },
      "KycSubmissionStatus": {
        "type": "string",
        "enum": [
          "PENDING",
          "APPROVED",
          "REJECTED"
        ]
      },
...
      "KycSubmissionDto": {
        "type": "object",
        "properties": {
          "targetTier": {
            "allOf": [
              {
                "$ref": "#/components/schemas/KycTier"
              }
            ]
          },
          "status": {
            "allOf": [
              {
                "$ref": "#/components/schemas/KycSubmissionStatus"
              }
            ]
          },
          "documentType": {
            "allOf": [
              {
                "$ref": "#/components/schemas/KycDocumentType"
              }
            ]
          },
          "documentRef": {
            "type": "string",
            "nullable": true
          },
          "businessName": {
            "type": "string",
            "nullable": true
          },
          "verificationResult": {
            "type": "object",
            "additionalProperties": true,
            "nullable": true
          },
          "reviewedById": {
            "type": "string",
            "nullable": true
          },
          "reviewedAt": {
            "type": "string",
            "nullable": true
          },
          "rejectionReason": {
            "type": "string",
            "nullable": true
          },
          "id": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "documentNumberLast4": {
            "type": "string"
          },
          "createdAt": {
            "type": "string"
          }
        },
        "required": [
          "targetTier",
          "status",
          "documentType",
          "documentRef",
          "businessName",
          "verificationResult",
          "reviewedById",
          "reviewedAt",
          "rejectionReason",
          "id",
          "userId",
          "documentNumberLast4",
          "createdAt"
        ]
      },
      "KycMeDto": {
        "type": "object",
        "properties": {
          "tier": {
            "allOf": [
              {
                "$ref": "#/components/schemas/KycTier"
              }
            ]
          },
          "limits": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TierLimitDto"
            }
          },
          "submissions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/KycSubmissionDto"
            }
          }
        },
        "required": [
          "tier",
          "limits",
          "submissions"
        ]
      },
      "SubmitKycDto": {
        "type": "object",
        "properties": {
          "targetTier": {
            "type": "string",
            "enum": [
              "TIER_1",
              "TIER_2",
              "TIER_3"
            ]
          },
          "documentType": {
            "type": "string",
            "enum": [
              "CNIC",
              "PASSPORT",
              "EMIRATES_ID",
              "DRIVING_LICENSE",
              "BUSINESS_REGISTRATION"
            ]
          },
          "documentNumber": {
            "type": "string",
            "minLength": 5,
            "maxLength": 40
          },
          "documentRef": {
            "type": "string",
            "description": "Reference to an uploaded document image in object storage.",
            "minLength": 1,
            "maxLength": 256
          },
          "dateOfBirth": {
            "type": "string",
            "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
          },
          "address": {
            "type": "string",
            "minLength": 5,
            "maxLength": 300
          },
          "businessName": {
            "type": "string",
            "minLength": 2,
            "maxLength": 200
          }
        },
        "required": [
          "targetTier",
          "documentType",
          "documentNumber"
        ]
      },
      "ApproveKycDto": {
        "type": "object",
        "properties": {
          "note": {
            "type": "string",
            "minLength": 1,
            "maxLength": 500
          }
        }
      },
      "RejectKycDto": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string",
            "minLength": 5,
            "maxLength": 500
          }
        },
        "required": [
          "reason"
        ]
      },
      "KycAuditLogDto": {
        "type": "object",
        "properties": {
          "submissionId": {
            "type": "string",
            "nullable": true
          },
          "actorId": {
            "type": "string",
            "nullable": true,
            "description": "null = SYSTEM (automated decision)"
          },
          "fromTier": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/KycTier"
              }
            ]
          },
          "toTier": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/KycTier"
              }
            ]
          },
          "details": {
            "type": "object",
            "additionalProperties": true,
            "nullable": true
          },
          "id": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "action": {
            "type": "string"
          },
          "createdAt": {
            "type": "string"
          }
        },
        "required": [
          "submissionId",
          "actorId",
          "fromTier",
          "toTier",
          "details",
          "id",
          "userId",
          "action",
          "createdAt"
        ]
      },
...
    }
  }
}